[ICU-4288] Fix some compiler warnings. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/17303 Former-commit-id: 28141b85c5adfeceb9c1fb1b857afc3cb5f1fb2c

+++ icu4c/source/common/rbbi.cpp
@@ -650,7 +650,7 @@ UBool RuleBasedBreakIterator::isBoundary(int32_t offset) {
  * @return The current iteration position.
  */
 int32_t RuleBasedBreakIterator::current(void) const {
-    return (fText != NULL) ? fText->getIndex() : BreakIterator::DONE;
+    return (fText != NULL) ? fText->getIndex() : (int32_t)BreakIterator::DONE;
 }
 
 //=======================================================================
